+++ extras/example.R
@@ -4,51 +4,70 @@ library(PatientLevelPrediction)
 library(DeepPatientLevelPrediction)
 
 data(plpDataSimulationProfile)
-sampleSize <- 1e4
+sampleSize <- 1e3
 plpData <- simulatePlpData(
-  plpDataSimulationProfile,
-  n = sampleSize 
-)
+   plpDataSimulationProfile,
+   n = sampleSize
+ )
+
 
 populationSet <- PatientLevelPrediction::createStudyPopulationSettings(
   requireTimeAtRisk = F, 
   riskWindowStart = 1, 
-  riskWindowEnd = 365)
+  riskWindowEnd = 365*5)
 
- 
-modelSettings <- setResNet(numLayers = 2, sizeHidden = 64, hiddenFactor = 1,
-                          residualDropout = 0, hiddenDropout = 0.2, normalization = 'BatchNorm',
-                          activation = 'RelU', sizeEmbedding = 512, weightDecay = 1e-6,
-                          learningRate = 3e-4, seed = 42, hyperParamSearch = 'random',
-                          randomSample = 1, device = 'cuda:0',batchSize = 32,epochs = 10)
+# 
+# modelSettings <- setDefaultTransformer(estimatorSettings = setEstimator(
+#   learningRate = "auto",
+#   batchSize=64L,
+#   epochs = 10L
+# ))
 
-# modelSettings <- setTransformer(numBlocks=1, dimToken = 33, dimOut = 1, numHeads = 3,
-#                                 attDropout = 0.2, ffnDropout = 0.2, resDropout = 0,
-#                                 dimHidden = 8, batchSize = 32, hyperParamSearch = 'random',
-#                                 weightDecay = 1e-6, learningRate = 3e-4, epochs = 10,
-#                                 device = 'cuda:0', randomSamples = 1, seed = 42)
+modelSettings <- setDefaultResNet(estimatorSettings = setEstimator(
+  learningRate = "auto",
+  weightDecay = 1e-06,
+  device="cuda:0",
+  batchSize=128L,
+  epochs=50L,
+  seed=42
+))
+
+modelSettings <- setResNet(numLayers = c(1L, 2L),
+                           sizeHidden = 72L,
+                           hiddenFactor = 1L,
+                           residualDropout = 0.0,
+                           hiddenDropout = 0.0,
+                           sizeEmbedding = 64L,
+                           estimatorSettings = setEstimator(
+                             learningRate = 3e-4,
+                             batchSize = 128L,
+                             epochs = 10L,
+                             device = "cpu",
+                             seed = 42
+                           ),
+                           randomSample = 2)
 
 res2 <- PatientLevelPrediction::runPlp(
-plpData = plpData,
-outcomeId = 3,
-modelSettings = modelSettings,
-analysisId = 'Test',
-analysisName = 'Testing DeepPlp',
-populationSettings = populationSet,
-splitSettings = createDefaultSplitSetting(),
-sampleSettings = createSampleSettings(),  # none
-featureEngineeringSettings = createFeatureEngineeringSettings(), # none
-preprocessSettings = createPreprocessSettings(),
-logSettings = createLogSettings(verbosity='TRACE'),
-executeSettings = createExecuteSettings(
-  runSplitData = T,
-  runSampleData = F,
-  runfeatureEngineering = F,
-  runPreprocessData = T,
-  runModelDevelopment = T,
-  runCovariateSummary = T
-),
-saveDirectory = '~/test/new_plp/'
+  plpData = plpData,
+  outcomeId = unique(plpData$outcomes$outcomeId)[[1]],
+  modelSettings = modelSettings,
+  analysisId = 'Test',
+  analysisName = 'Testing DeepPlp',
+  populationSettings = populationSet,
+  splitSettings = createDefaultSplitSetting(splitSeed = 123),
+  sampleSettings = createSampleSettings("underSample"),  # none
+  featureEngineeringSettings = createFeatureEngineeringSettings(), # none
+  preprocessSettings = createPreprocessSettings(normalize = F),
+  logSettings = createLogSettings(verbosity='TRACE'),
+  executeSettings = createExecuteSettings(
+    runSplitData = T,
+    runSampleData = T,
+    runfeatureEngineering = F,
+    runPreprocessData = T,
+    runModelDevelopment = T,
+    runCovariateSummary = F
+  ),
+  saveDirectory = '~/test/resnet/'
 )