+++ tests/testthat/test-TrainingCache.R
@@ -1,17 +1,17 @@
 resNetSettings <- setResNet(numLayers = c(1, 2, 4),
-                            sizeHidden = 2^6,
+                            sizeHidden = 64,
                             hiddenFactor = 1,
                             residualDropout = 0.5,
                             hiddenDropout = 0.5,
-                            sizeEmbedding = 2^6,
-                            estimatorSettings = setEstimator(learningRate='auto',
+                            sizeEmbedding = 64,
+                            estimatorSettings = setEstimator(learningRate=3e-4,
                                                              weightDecay=1e-3,
                                                              device='cpu',
-                                                             batchSize=1024,
-                                                             epochs=30,
+                                                             batchSize=64,
+                                                             epochs=1,
                                                              seed=NULL),
                             hyperParamSearch = "random",
-                            randomSample = 2,
+                            randomSample = 3,
                             randomSampleSeed = NULL)
 
 trainCache <- TrainingCache$new(testLoc)
@@ -84,4 +84,6 @@ test_that("Estimator can resume training from cache", {
     }
   )
   sink()
+  trainCache <- TrainingCache$new(analysisPath)
+  testthat::expect_equal(is.na(trainCache$getLastGridSearchIndex()), TRUE)
 })