+++ tests/testthat/test-Transformer.R
@@ -1,7 +1,12 @@
 settings <- setTransformer(
-  numBlocks = 1, dimToken = 8, dimOut = 1,
-  numHeads = 2, attDropout = 0.0, ffnDropout = 0.2,
-  resDropout = 0.0, dimHidden = 32, 
+  numBlocks = 1, 
+  dimToken = 8, 
+  dimOut = 1,
+  numHeads = 2, 
+  attDropout = 0.0, 
+  ffnDropout = 0.2,
+  resDropout = 0.0, 
+  dimHidden = 32, 
   estimatorSettings = setEstimator(learningRate = 3e-4,
                                    batchSize=64,
                                    epochs=1),
@@ -39,35 +44,48 @@ test_that("fitEstimator with Transformer works", {
 })
 
 test_that("transformer nn-module works", {
+  Transformer <- reticulate::import_from_path("Transformer", path=path)$Transformer
   model <- Transformer(
-    catFeatures = 5, numFeatures = 1, numBlocks = 2,
-    dimToken = 16, numHeads = 2, attDropout = 0, ffnDropout = 0,
-    resDropout = 0, dimHidden = 32
+    cat_features = 5, 
+    num_features = 1, 
+    num_blocks = 2,
+    dim_token = 16, 
+    num_heads = 2, 
+    att_dropout = 0, 
+    ffn_dropout = 0,
+    res_dropout = 0, 
+    dim_hidden = 32
   )
 
-  pars <- sum(sapply(model$parameters, function(x) prod(x$shape)))
+  pars <- sum(reticulate::iterate(model$parameters(), function(x) x$numel()))
 
   # expected number of parameters
   expect_equal(pars, 5697)
 
   input <- list()
-  input$cat <- torch::torch_randint(0, 5, c(10, 5), dtype = torch::torch_long())
-  input$num <- torch::torch_randn(10, 1, dtype = torch::torch_float32())
+  input$cat <- torch$randint(0L, 5L, c(10L, 5L), dtype = torch$long)
+  input$num <- torch$randn(10L, 1L, dtype = torch$float32)
 
   output <- model(input)
 
   # output is correct shape, size of batch
-  expect_equal(output$shape, 10)
+  expect_equal(output$shape[0], 10L)
 
   input$num <- NULL
 
   model <- Transformer(
-    catFeatures = 5, numFeatures = 0, numBlocks = 2,
-    dimToken = 16, numHeads = 2, attDropout = 0, ffnDropout = 0,
-    resDropout = 0, dimHidden = 32
+    cat_features = 5, 
+    num_features = 0, 
+    num_blocks = 2,
+    dim_token = 16, 
+    num_heads = 2, 
+    att_dropout = 0, 
+    ffn_dropout = 0,
+    res_dropout = 0, 
+    dim_hidden = 32
   )
   output <- model(input)
-  expect_equal(output$shape, 10)
+  expect_equal(output$shape[0], 10L)
 })
 
 test_that("Default Transformer works", {
@@ -89,4 +107,22 @@ test_that("Errors are produced by settings function", {
   randomSample <- 2
   
   expect_error(setTransformer(randomSample = randomSample))
+})
+
+test_that("dimHidden ratio works as expected", {
+  randomSample <- 4
+  dimToken <- c(64, 128, 256, 512)
+  dimHiddenRatio <- 2
+  modelSettings <- setTransformer(dimToken = dimToken,
+                                  dimHiddenRatio = dimHiddenRatio,
+                                  dimHidden = NULL,
+                                  randomSample = randomSample)
+  dimHidden <- unlist(lapply(modelSettings$param, function(x) x$dimHidden))
+  tokens <-   unlist(lapply(modelSettings$param, function(x) x$dimToken))
+  testthat::expect_true(all(dimHidden == dimHiddenRatio * tokens))
+  testthat::expect_error(setTransformer(dimHidden = NULL,
+                                        dimHiddenRatio = NULL))
+  testthat::expect_error(setTransformer(dimHidden = 256,
+                                        dimHiddenRatio = 4/3))
+
 })